requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.26.5
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
pycodestyle==2.10.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.28.2
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
